@@ -8,22 +8,24 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@ionic/core": "5.3.0",
+    "@modus/ionic-vue": "^3.0.0-alpha.13",
     "core-js": "^3.6.5",
     "register-service-worker": "^1.7.1",
-    "vue": "^2.6.11",
-    "vue-router": "^3.2.0",
-    "vuex": "^3.4.0"
+    "vue": "3.0.0-rc.4",
+    "vue-router": "4.0.0-beta.5",
+    "vuex": "^4.0.0-alpha.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.4.0",
-    "@vue/cli-plugin-eslint": "~4.4.0",
     "@vue/cli-plugin-pwa": "~4.4.0",
     "@vue/cli-plugin-router": "~4.4.0",
     "@vue/cli-plugin-vuex": "~4.4.0",
     "@vue/cli-service": "~4.4.0",
+    "@vue/compiler-sfc": "3.0.0-rc.4",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0-alpha.0",
+    "vue-cli-plugin-vue-next": "~0.1.3"
   }
 }
